+++ crates/ruma-client-api/src/redact/redact_event.rs
@@ -8,7 +8,7 @@ pub mod v3 {
     //! [spec]: https://spec.matrix.org/v1.19/client-server-api/#put_matrixclientv3roomsroomidredacteventidtxnid
 
     use ruma_common::{
-        EventId, OwnedTransactionId, RoomId,
+        EventId, RoomId, TransactionId,
         api::{auth_scheme::AccessToken, request, response},
         metadata,
     };
@@ -44,7 +44,7 @@ pub mod v3 {
         ///
         /// [access token is refreshed]: https://spec.matrix.org/v1.19/client-server-api/#refreshing-access-tokens
         #[ruma_api(path)]
-        pub txn_id: OwnedTransactionId,
+        pub txn_id: TransactionId,
 
         /// The reason for the redaction.
         #[serde(skip_serializing_if = "Option::is_none")]
@@ -60,7 +60,7 @@ pub mod v3 {
 
     impl Request {
         /// Creates a new `Request` with the given room ID, event ID and transaction ID.
-        pub fn new(room_id: RoomId, event_id: EventId, txn_id: OwnedTransactionId) -> Self {
+        pub fn new(room_id: RoomId, event_id: EventId, txn_id: TransactionId) -> Self {
             Self { room_id, event_id, txn_id, reason: None }
         }
     }